feat: User 권한 정책 구현 (#100)
* feat: user 조회 api 인가 정책 추가 : admin만 조회 가능

* feat: 개인정보 조회를 위한 `POST /users/me` api 추가

Response를 User에서 가공을 하여 보여주는 부분을 수정할 필요가 있음.

* feat:  유저 삭제 권한 추가 및 `DELETE /users/me` api 추가

* feat: 유저 자신의 패쓰워드 수정 기능 구현

`PUT /users/me/password`
SignUpRequest의 유효성 검증 부분 추가

* feat: 유저 권한 수정 기능 추가

admin, Admin, ADMIN 과 같이 다양한 요청에 대해서도 기능이 동작해야 한다고 판단하여 `ignoringCaseValueOf()`메소드를 구현하여 해결

* feat: 자신의 계정 삭제시 패쓰워드 필요

+++ src/main/java/com/example/gistcompetitioncnserver/user/UserController.java
@@ -1,5 +1,6 @@
 package com.example.gistcompetitioncnserver.user;
 
+import com.example.gistcompetitioncnserver.exception.CustomException;
 import lombok.AllArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.http.ResponseEntity;
@@ -41,17 +42,81 @@ public class UserController {
 
     @GetMapping("/users")
     public ResponseEntity<List<User>> retrieveAllUsers() {
+        SessionUser sessionUser = (SessionUser) httpSession.getAttribute("user");
+        if (!sessionUser.getEnabled()) {
+            throw new CustomException("이메일 인증이 필요합니다!");
+        }
+        if (!sessionUser.isAdmin()) {
+            throw new CustomException("회원 정보 조회 권한이 없습니다.");
+        }
         return ResponseEntity.ok().body(userService.findAllUsers());
     }
 
     @GetMapping("/users/{userId}")
     public ResponseEntity<User> retrieveUser(@PathVariable Long userId) {
+        SessionUser sessionUser = (SessionUser) httpSession.getAttribute("user");
+        if (!sessionUser.getEnabled()) {
+            throw new CustomException("이메일 인증이 필요합니다!");
+        }
+        if (!sessionUser.isAdmin()) {
+            throw new CustomException("회원 정보 조회 권한이 없습니다.");
+        }
         return ResponseEntity.ok().body(userService.findUserById(userId));
+    }
+
+    @GetMapping("/users/me")
+    public ResponseEntity<User> retrieveUserOfMine() {
+        SessionUser sessionUser = (SessionUser) httpSession.getAttribute("user");
+        if (!sessionUser.getEnabled()) {
+            throw new CustomException("이메일 인증이 필요합니다!");
+        }
+        return ResponseEntity.ok().body(userService.findUserById(sessionUser.getId()));
+    }
+
+    @PutMapping("/users/{userId}/userRole")
+    public ResponseEntity<Void> updateUserRole(@PathVariable Long userId,
+                                               @Validated @RequestBody UpdateUserRoleRequest userRoleRequest) {
+        SessionUser sessionUser = (SessionUser) httpSession.getAttribute("user");
+        if (!sessionUser.getEnabled()) {
+            throw new CustomException("이메일 인증이 필요합니다!");
+        }
+        if (!sessionUser.isAdmin()) {
+            throw new CustomException("유저 권한을 수정할 권한이 없습니다.");
+        }
+        userService.updateUserRole(userId, userRoleRequest);
+        return ResponseEntity.noContent().build();
+    }
+
+    @PutMapping("/users/me/password")
+    public ResponseEntity<Void> updatePasswordOfMine(@Validated @RequestBody UpdatePasswordRequest passwordRequest) {
+        SessionUser sessionUser = (SessionUser) httpSession.getAttribute("user");
+        if (!sessionUser.getEnabled()) {
+            throw new CustomException("이메일 인증이 필요합니다!");
+        }
+        userService.updatePassword(sessionUser.getId(), passwordRequest);
+        return ResponseEntity.noContent().build();
     }
 
     @DeleteMapping("/users/{userId}")
     public ResponseEntity<Void> deleteUser(@PathVariable Long userId) {
+        SessionUser sessionUser = (SessionUser) httpSession.getAttribute("user");
+        if (!sessionUser.getEnabled()) {
+            throw new CustomException("이메일 인증이 필요합니다!");
+        }
+        if (!sessionUser.isAdmin()) {
+            throw new CustomException("삭제할 권한이 없습니다.");
+        }
         userService.deleteUser(userId);
+        return ResponseEntity.noContent().build();
+    }
+
+    @DeleteMapping("/users/me")
+    public ResponseEntity<Void> deleteUserOfMine(@Validated @RequestBody DeleteUserRequest deleteUserRequest) {
+        SessionUser sessionUser = (SessionUser) httpSession.getAttribute("user");
+        if (!sessionUser.getEnabled()) {
+            throw new CustomException("이메일 인증이 필요합니다!");
+        }
+        userService.deleteUserOfMine(sessionUser.getId(), deleteUserRequest);
         return ResponseEntity.noContent().build();
     }
 }
